removing users from the url and added a cards route
