+++ app/routes/decks.js
@@ -1,13 +1,6 @@
 import Ember from 'ember';
-import Card from '../models/card';
-import Set from '../models/set';
-import ENV from '../config/environment';
-
-function _createCards (cardKeys, data) {
-  return cardKeys.map(function (cardKey) {
-    return Card.create(data[cardKey]);
-  });
-}
+import Set from 'webatrice/models/set';
+import ENV from 'webatrice/config/environment';
 
 function _createSets (sets) {
   return sets.map(function (s) {
@@ -16,24 +9,6 @@ function _createSets (sets) {
 }
 
 export default Ember.Route.extend({
-  beforeModel: function () {
-    var cardsController = this.controllerFor('cards');
-    if (ENV.environment === 'development') {
-      return Ember.$.getJSON('/api/cards').then(function (data) {
-        cardsController.set('model', _createCards(Ember.keys(data.cards[0]), data.cards[0]));
-      });
-    } else {
-      return Ember.$.ajax({
-        url: 'http://mtgjson.com/json/AllCards-x.jsonp',
-        jsonpCallback: 'mtgjsoncallback',
-        dataType: 'jsonp',
-        success: function (data) {
-          cardsController.set('model', _createCards(Ember.keys(data), data));
-        }
-      });
-    }
-  },
-
   model: function () {
     return [];
   },
