+++ app/templates/application.hbs
@@ -12,11 +12,15 @@
           </ul>
         </li>
         <li class="dropdown">
+          {{#if isLoggedIn}}
           <a class="dropdown-toggle" data-toggle="dropdown">Hi {{controllers.user.name}}{{fa-icon "caret-down"}}</a>
           <ul class="dropdown-menu">
             <li>{{#link-to 'user' controllers.user.id}}Profile{{/link-to}}</li>
             <li>{{#link-to 'decks.list' controllers.user.id}}Decks{{/link-to}}</li>
           </ul>
+          {{else}}
+            {{#link-to 'login'}}Login{{/link-to}}
+          {{/if}}
         </li>
       </ul>
     </div><!--/.nav-collapse -->
