+++ app/templates/cards.hbs
@@ -1,0 +1,3 @@
+{{#full-height class="overflow-y-scroll"}}
+  {{partial "cards/table-of-cards"}}
+{{/full-height}}
